challenger/requirements.txt
fastapi
uvicorn[standard]
markitdown[all]
pillow
python-multipart
requests
reportlab